Backend/FinallApp/requirements.txt
streamlit
pandas
openpyxl
mysql-connector-python
pdfplumber
pywin32
xlrd
fuzzywuzzy
python-Levenshtein
